refactor: move test helpers to dedicated file and update imports
